+++ tests/peasant_test.rs
@@ -8,9 +8,9 @@ mod test {
 
   #[test]
   fn test_peasant() {
-    fn peasant_is_as_multiplication(a: u32, b:u32) -> bool {
+    fn peasant_is_as_multiplication(a: usize, b:usize) -> bool {
       a.checked_mul(b) == peasant::peasant(a, b)
     }
-    quickcheck(peasant_is_as_multiplication as fn(u32, u32) -> bool)
+    quickcheck(peasant_is_as_multiplication as fn(usize, usize) -> bool)
   }  
 }
